+++ mols2grid/__init__.py
@@ -2,8 +2,14 @@ from .molgrid import MolGrid
 from .dispatch import (display,
                        save)
 from .select import (get_selection,
-                     selection,
                      list_grids)
 from .utils import (sdf_to_dataframe,
                     make_popup_callback)
 from ._version import __version__
+
+try:
+    from google.colab import output
+except ImportError:
+    pass
+else:
+    output.enable_custom_widget_manager()